Hacky way to paper over a bug: cluster deletion
    107 21 CreationTimestamp 1264022981
    101 1.0 Job Machine
    103 1.0 B 2
    101 01.-1 Job Machine
    103 01.-1 A 1
    102 1.0
    102 01.-1

The above example deletes both 1.0 and its parent (01.-1). A Job holds
a pointer to its parent Job. The deletion loop happens to be sorted in
such a way that the parent Jobs are deleted first, which means the
child's parent pointer is garbage when it comes around to use it.

This commit papers over the bug by reversing the deletion order to
delete children first. A proper solution would delay the parent
deletion until all children have been deleted, without relying on the
sorted order of the g_delete_jobs set.